Moved tilesize to only exist in the draw methods for entities Changed the representaiton of a position from a Point to a Point2D.Double to be able to be "between" tiles.

+++ src/se/liu/antos931jakos322/towerdefence/other/HelperFunctions.java
@@ -13,7 +13,7 @@ package se.liu.antos931jakos322.towerdefence.other;
 public class HelperFunctions
 {
 
-    public static double pythagoras(int a, int b){
+    public static double pythagoras(double a, double b){
 	// returns the integer value of pythagoras therom for 2 integers
 	double c = Math.pow((a), 2) + Math.pow((b), 2);
